+++ packages/@ionic/cli-utils/src/lib/app.ts
@@ -1,8 +1,8 @@
 import chalk from 'chalk';
 
-import { App, IClient, IPaginator, ResourceClient, Response } from '../definitions';
-import { isAppResponse, isAppsResponse } from '../guards';
-import { createFatalAPIFormat } from './http';
+import { App, AppAssociation, IClient, IPaginator, PaginateArgs, PaginatorState, ResourceClientCreate, ResourceClientLoad, ResourceClientPaginate, Response } from '../definitions';
+import { isAppAssociationResponse, isAppResponse, isAppsResponse } from '../guards';
+import { ResourceClient, createFatalAPIFormat } from './http';
 
 export function formatName(app: Pick<App, 'name' | 'org'>) {
   if (app.org) {
@@ -21,18 +21,19 @@ export interface AppCreateDetails {
   name: string;
 }
 
-export class AppClient implements ResourceClient<App, AppCreateDetails> {
+export class AppClient extends ResourceClient implements ResourceClientLoad<App>, ResourceClientCreate<App, AppCreateDetails>, ResourceClientPaginate<App> {
   protected client: IClient;
   protected token: string;
 
   constructor({ client, token }: AppClientDeps) {
+    super();
     this.client = client;
     this.token = token;
   }
 
   async load(id: string): Promise<App> {
     const { req } = await this.client.make('GET', `/apps/${id}`);
-    req.set('Authorization', `Bearer ${this.token}`);
+    this.applyAuthentication(req, this.token);
     const res = await this.client.do(req);
 
     if (!isAppResponse(res)) {
@@ -44,7 +45,8 @@ export class AppClient implements ResourceClient<App, AppCreateDetails> {
 
   async create({ name }: AppCreateDetails): Promise<App> {
     const { req } = await this.client.make('POST', '/apps');
-    req.set('Authorization', `Bearer ${this.token}`).send({ name });
+    this.applyAuthentication(req, this.token);
+    req.send({ name });
     const res = await this.client.do(req);
 
     if (!isAppResponse(res)) {
@@ -54,14 +56,41 @@ export class AppClient implements ResourceClient<App, AppCreateDetails> {
     return res.data;
   }
 
-  paginate(): IPaginator<Response<App[]>> {
-    return this.client.paginate(
-      async () => {
+  paginate(args: Partial<PaginateArgs<Response<App[]>>> = {}): IPaginator<Response<App[]>, PaginatorState> {
+    return this.client.paginate({
+      reqgen: async () => {
         const { req } = await this.client.make('GET', '/apps');
-        req.set('Authorization', `Bearer ${this.token}`);
+        this.applyAuthentication(req, this.token);
         return { req };
       },
-      isAppsResponse
-    );
+      guard: isAppsResponse,
+      ...args,
+    });
+  }
+
+  async createGithubAssociation(id: string, association: { repoId: string; }): Promise<AppAssociation> {
+    const { req } = await this.client.make('POST', `/apps/${id}/github`);
+
+    req
+      .set('Authorization', `Bearer ${this.token}`)
+      .send({ repository_id: association.repoId });
+
+    const res = await this.client.do(req);
+
+    if (!isAppAssociationResponse(res)) {
+      throw createFatalAPIFormat(req, res);
+    }
+
+    return res.data;
+  }
+
+  async deleteGithubAssociation(id: string): Promise<void> {
+    const { req } = await this.client.make('DELETE', `/apps/${id}/github`);
+
+    req
+      .set('Authorization', `Bearer ${this.token}`)
+      .send({});
+
+    await req;
   }
 }