+++ packages/@ionic/cli-utils/src/lib/errors.ts
@@ -1,21 +1,6 @@
-export class Exception extends Error {
-  message: string;
-  name: string;
-  stack: string;
+import { BaseError } from '@ionic/cli-framework/lib/errors';
 
-  constructor(message: string) {
-    super(message);
-    this.name = 'Exception';
-    this.message = message;
-    this.stack = (new Error()).stack || '';
-  }
-
-  toString() {
-    return `${this.name}: ${this.message}`;
-  }
-}
-
-export class FatalException extends Exception {
+export class FatalException extends BaseError {
   fatal = true;
 
   constructor(public message = '', public exitCode = 1) {
@@ -23,24 +8,24 @@ export class FatalException extends Exception {
   }
 }
 
-export class ShellException extends Exception {
+export class ShellException extends BaseError {
   constructor(public message: string, public exitCode = 0) {
     super(message);
   }
 }
 
-export class ServeException extends Exception {}
+export class ServeException extends BaseError {}
 
 export class ServeCommandNotFoundException extends ServeException {}
 
-export class SessionException extends Exception {}
+export class SessionException extends BaseError {}
 
-export class RunnerException extends Exception {}
+export class RunnerException extends BaseError {}
 
 export class RunnerNotFoundException extends RunnerException {}
 
-export class IntegrationException extends Exception {}
+export class IntegrationException extends BaseError {}
 
 export class IntegrationNotFoundException extends IntegrationException {}
 
-export class HookException extends Exception {}
+export class HookException extends BaseError {}