+++ packages/@ionic/cli-utils/src/lib/security.ts
@@ -1,24 +1,26 @@
-import { IClient, ResourceClient, SecurityProfile } from '../definitions';
+import { IClient, ResourceClientLoad, SecurityProfile } from '../definitions';
 import { isSecurityProfileResponse } from '../guards';
-import { createFatalAPIFormat } from './http';
+import { ResourceClient, createFatalAPIFormat } from './http';
 
 export interface SecurityClientDeps {
   readonly client: IClient;
   readonly token: string;
 }
 
-export class SecurityClient implements ResourceClient<SecurityProfile, never> {
+export class SecurityClient extends ResourceClient implements ResourceClientLoad<SecurityProfile> {
   protected client: IClient;
   protected token: string;
 
   constructor({ client, token }: SecurityClientDeps) {
+    super();
     this.client = client;
     this.token = token;
   }
 
   async load(tag: string): Promise<SecurityProfile> {
     const { req } = await this.client.make('GET', `/security/profiles/${tag}`);
-    req.set('Authorization', `Bearer ${this.token}`).query({}).send();
+    this.applyAuthentication(req, this.token);
+    req.query({}).send();
     const res = await this.client.do(req);
 
     if (!isSecurityProfileResponse(res)) {