+++ packages/ionic/src/commands/integrations/disable.ts
@@ -1,9 +1,10 @@
 import chalk from 'chalk';
 
 import { contains, validators } from '@ionic/cli-framework';
+import { BaseError } from '@ionic/cli-framework/lib/errors';
 import { CommandLineInputs, CommandLineOptions, CommandMetadata, isIntegrationName } from '@ionic/cli-utils';
 import { Command } from '@ionic/cli-utils/lib/command';
-import { Exception, FatalException } from '@ionic/cli-utils/lib/errors';
+import { FatalException } from '@ionic/cli-utils/lib/errors';
 import { BaseIntegration, INTEGRATION_NAMES } from '@ionic/cli-utils/lib/integrations';
 
 export class IntegrationsDisableCommand extends Command {
@@ -40,7 +41,7 @@ export class IntegrationsDisableCommand extends Command {
         this.env.log.ok(`Integration ${chalk.green(integration.name)} disabled!`);
       }
     } catch (e) {
-      if (e instanceof Exception) {
+      if (e instanceof BaseError) {
         throw new FatalException(e.message);
       }